Refactor: move factories and launcher in bootstrap/
