+++ test.py
@@ -9,8 +9,8 @@
 from hydra_zen import store, zen
 
 import conf.experiment  # Must import the config to add all components to the store! # noqa
+from bootstrap.launch_experiment import launch_experiment
 from conf import project as project_conf
-from launch_experiment import launch_experiment
 from utils import seed_everything
 
 if __name__ == "__main__":
